Updated a bunch of random debugging stuff
