+++ app/src/main/res/menu/menu_main.xml
@@ -1,9 +1,35 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto"
-      xmlns:tools="http://schemas.android.com/tools"
-      tools:context=".MainActivity">
-    <item android:id="@+id/action_settings"
-          android:title="@string/action_settings"
-          android:orderInCategory="100"
-          app:showAsAction="never"/>
+<menu
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".MainActivity">
+
+    <item
+        android:id="@+id/action_enable_rotation"
+        android:checkable="true"
+        android:checked="true"
+        android:title="@string/action_enable_rotation"
+        app:showAsAction="never"/>
+
+    <item
+        android:id="@+id/action_show_control_points"
+        android:checkable="true"
+        android:checked="false"
+        android:title="@string/action_show_control_points"
+        app:showAsAction="never"/>
+
+    <item
+        android:id="@+id/action_show_bounds"
+        android:checkable="true"
+        android:checked="false"
+        android:title="@string/action_show_bounds"
+        app:showAsAction="never"/>
+
+    <item
+        android:id="@+id/action_slow_down_animation"
+        android:checkable="true"
+        android:checked="false"
+        android:title="@string/action_slow_down_animation"
+        app:showAsAction="never"/>
+
 </menu>
