+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,14 @@
 <resources>
-    <string name="app_name">Check Mark Animation</string>
 
-    <string name="hello_world">Hello world!</string>
-    <string name="action_settings">Settings</string>
+    <string name="app_name">Bezier Icon Transformation</string>
+
+    <string name="refresh">Refresh</string>
+    <string name="check">Check</string>
+    <string name="exclamation">Exclamation</string>
+
+    <string name="action_enable_rotation">Show rotation</string>
+    <string name="action_show_control_points">Show control points</string>
+    <string name="action_show_bounds">Show debug bounds</string>
+    <string name="action_slow_down_animation">Slow animation</string>
+
 </resources>
