--- app/src/main/res/values/styles.xml
@@ -1,7 +1,0 @@
-<resources>
-
-    <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
-        <item name="android:statusBarColor">#000</item>
-    </style>
-
-</resources>
